pyproject.toml
[tool.poetry]
name = "aws-iot-core-app"
version = "0.0.0"
description = ""
authors = ["Bhaktij Koli <bhaktijkoli121@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
awsiotsdk = "^1.11.3"
loguru = "^0.6.0"
fastapi = {extras = ["all"], version = "^0.79.0"}
uuid = "^1.30"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
